#164 move inspection to the Code Style group. Covered break and continue situations
